pangkat desimal logaritma

+++ header.h
@@ -82,8 +82,8 @@ double to_radian(double degree);
 
 
 //Ahmad
-int loga(int a);
-int log_a_to_base_b(int a, int b);
+float loga(int a);
+float log_a_to_base_b(int a, int b);
 
 
 #endif
